# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Together API",
    "description": "Private, non-federated Discord alternative for small gaming communities",
    "contact": {
      "name": "Together Contributors"
    },
    "license": {
      "name": "PolyForm Noncommercial 1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/servers/{id}/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "GET /servers/:id/channels — list all channels in a server (members only).",
        "operationId": "list_channels",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Channel"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a server member"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "POST /servers/:id/channels — create a channel in a server (owner only).",
        "operationId": "create_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Channel created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Not the server owner"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/servers/{id}/channels/{channel_id}": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "GET /servers/:id/channels/:channel_id — get a single channel (members only).",
        "operationId": "get_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "403": {
            "description": "Not a server member"
          },
          "404": {
            "description": "Server or channel not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Channels"
        ],
        "summary": "DELETE /servers/:id/channels/:channel_id — delete a channel (owner only).",
        "description": "This is a hard delete; any messages in the channel are also removed by the\ndatabase cascade constraint (ON DELETE CASCADE on messages.channel_id).",
        "operationId": "delete_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel deleted"
          },
          "403": {
            "description": "Not the server owner"
          },
          "404": {
            "description": "Server or channel not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Channels"
        ],
        "summary": "PATCH /servers/:id/channels/:channel_id — update a channel (owner only).",
        "operationId": "update_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Channel updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Not the server owner"
          },
          "404": {
            "description": "Server or channel not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "type": "object",
        "required": [
          "id",
          "server_id",
          "name",
          "type",
          "position",
          "created_at"
        ],
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "server_id": {
            "type": "string",
            "format": "uuid"
          },
          "topic": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/ChannelType"
          }
        }
      },
...
      "ChannelType": {
        "type": "string",
        "enum": [
          "text",
          "voice"
        ]
      },
...
      "CreateChannelRequest": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "topic": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/ChannelType"
          }
        }
      },
...
      "UpdateChannelRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "topic": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}